refactor(editor): extract BOM row builders to state/bom.ts
context.svelte.ts had ~100 lines of BOM-specific helpers
(formatLength, nodeDisplayLabel, endpointRequirementKey,
cableRequirementKeys, buildAssignmentRows) mixed in with the
catalog/PoE/cross-store glue. Move them out to lib/state/bom.ts
where they live by themselves and accept their dependencies as
explicit args (nodes, links, scenes) instead of reaching into
module-level stores.

context now keeps only the thin wrapper that derives the rows
from the live stores.
